WIP: WIP

+++ src/backend/models/GetShiftedImagesModel.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel
 
 
 class GetShiftedImagesModel(BaseModel):
-    z: list[int]
+    z: list[float]
     shifts_range: int
     shifts_count: int
     dim: int
